Add jaeger tracing feature.
