+++ pkg/tracing/gcloud/gcloud.go
@@ -1,0 +1,131 @@
+package gcloud
+
+import (
+	"context"
+	"github.com/prometheus/common/version"
+	"os"
+
+	"fmt"
+
+	"cloud.google.com/go/trace/apiv1"
+	"github.com/go-kit/kit/log"
+	"github.com/go-kit/kit/log/level"
+	"github.com/lovoo/gcloud-opentracing"
+	"github.com/opentracing/basictracer-go"
+	"github.com/opentracing/opentracing-go"
+)
+
+const ForceTracingBaggageKey = "X-Thanos-Force-Tracing"
+
+type tracer struct {
+	debugName string
+	wrapped   opentracing.Tracer
+}
+
+func (t *tracer) StartSpan(operationName string, opts ...opentracing.StartSpanOption) opentracing.Span {
+	span := t.wrapped.StartSpan(operationName, opts...)
+
+	if t.debugName != "" {
+		span.SetTag("service_name", t.debugName)
+	}
+
+	// Set common tags.
+	if hostname := os.Getenv("HOSTNAME"); hostname != "" {
+		span.SetTag("hostname", hostname)
+	}
+
+	span.SetTag("binary_revision", version.Revision)
+	if len(os.Args) > 1 {
+		span.SetTag("binary_cmd", os.Args[1])
+	}
+
+	return span
+}
+
+func (t *tracer) Extract(format interface{}, carrier interface{}) (opentracing.SpanContext, error) {
+	return t.wrapped.Extract(format, carrier)
+}
+
+func (t *tracer) Inject(sm opentracing.SpanContext, format interface{}, carrier interface{}) error {
+	return t.wrapped.Inject(sm, format, carrier)
+}
+
+type forceRecorder struct {
+	wrapped basictracer.SpanRecorder
+}
+
+// RecordSpan invokes wrapper SpanRecorder only if Sampled field is true or ForceTracingBaggageKey item is set in span's context.
+// NOTE(bplotka): Currently only HTTP supports ForceTracingBaggageKey injection on ForceTracingBaggageKey header existence.
+func (r *forceRecorder) RecordSpan(sp basictracer.RawSpan) {
+	if force := sp.Context.Baggage[ForceTracingBaggageKey]; force != "" {
+		sp.Context.Sampled = true
+	}
+
+	// All recorder implementation should support handling sp.Context.Sampled.
+	r.wrapped.RecordSpan(sp)
+}
+
+
+type gcloudRecorderLogger struct {
+	logger log.Logger
+}
+
+func (l *gcloudRecorderLogger) Infof(format string, args ...interface{}) {
+	level.Info(l.logger).Log("msg", fmt.Sprintf(format, args...))
+}
+
+func (l *gcloudRecorderLogger) Errorf(format string, args ...interface{}) {
+	level.Error(l.logger).Log("msg", fmt.Sprintf(format, args...))
+}
+
+// NewOptionalGCloudTracer returns GoogleCloudTracer Tracer. In case of error it log warning and returns noop tracer.
+func NewOptionalGCloudTracer(ctx context.Context, logger log.Logger, gcloudTraceProjectID string, sampleFactor uint64, debugName string) (opentracing.Tracer, func() error) {
+	if gcloudTraceProjectID == "" {
+		level.Warn(logger).Log("msg", "gcloudtrace.project is empty. Google Cloud Tracer. Tracing will be disabled")
+		return &opentracing.NoopTracer{}, func() error { return nil }
+	}
+
+	tracer, closeFn, err := newGCloudTracer(ctx, logger, gcloudTraceProjectID, sampleFactor, debugName)
+	if err != nil {
+		level.Warn(logger).Log("msg", "failed to init Google Cloud Tracer. Tracing will be disabled", "err", err)
+		return &opentracing.NoopTracer{}, func() error { return nil }
+	}
+
+	level.Info(logger).Log("msg", "initiated Google Cloud Tracer. Tracing will be enabled", "err", err)
+	return tracer, closeFn
+}
+
+func newGCloudTracer(ctx context.Context, logger log.Logger, gcloudTraceProjectID string, sampleFactor uint64, debugName string) (opentracing.Tracer, func() error, error) {
+	traceClient, err := trace.NewClient(ctx)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	r, err := gcloudtracer.NewRecorder(
+		ctx,
+		gcloudTraceProjectID,
+		traceClient,
+		gcloudtracer.WithLogger(&gcloudRecorderLogger{logger: logger}))
+	if err != nil {
+		return nil, traceClient.Close, err
+	}
+
+	shouldSample := func(traceID uint64) bool {
+		// Set the sampling rate.
+		return traceID%sampleFactor == 0
+	}
+	if sampleFactor < 1 {
+		level.Debug(logger).Log("msg", "Tracing is enabled, but sampling is 0 which means only spans with 'force tracing' baggage will enable tracing.")
+		shouldSample = func(_ uint64) bool {
+			return false
+		}
+	}
+	return &tracer{
+		debugName: debugName,
+		wrapped: basictracer.NewWithOptions(basictracer.Options{
+			ShouldSample:   shouldSample,
+			Recorder:       &forceRecorder{wrapped: r},
+			MaxLogsPerSpan: 100,
+		}),
+	}, r.Close, nil
+}
